Move tests to convert.rs add checks for converting
Move tests from config_file.rs to convert.rs
and add checks for converting using `to_shell_source`

Uses `maplit` for the `hashmap!` macro.
