+++ Cargo.toml
@@ -30,6 +30,7 @@ env_logger = "0.9.0"
 
 [dev-dependencies]
 indoc = "1.0.4"
+maplit = "1.0.2"
 pretty_assertions = "1.2.1"
 
 [build-dependencies]
